+++ elements/storytelling/src/main.js
@@ -11,6 +11,7 @@ import {
   validateMarkdownAttrs,
   addLightBoxScript,
   addCustomSection,
+  initSavedMarkdown,
 } from "./helpers";
 import mainStyle from "../../../utils/styles/dist/main.style";
 import DOMPurify from "isomorphic-dompurify";
@@ -24,7 +25,11 @@ import { DEFAULT_SENSITIVE_TAGS, SAMPLE_ELEMENTS } from "./enums";
 const md = markdownit({ html: true });
 
 md.use(markdownItDecorateImproved).use(markdownItConfig);
-
+/**
+ * Manage drawn features on a map
+ *
+ * @element eox-storytelling
+ */
 export class EOxStoryTelling extends LitElement {
   // Define properties with defaults and types
   static get properties() {
@@ -33,7 +38,7 @@ export class EOxStoryTelling extends LitElement {
       markdownURL: { attribute: "markdown-url", type: String },
       nav: { state: true, attribute: false, type: Array },
       showNav: { attribute: "show-nav", type: Boolean },
-      showEditor: { attribute: "show-editor", type: Boolean },
+      showEditor: { attribute: "show-editor", type: String },
       noShadow: { attribute: "no-shadow", type: Boolean },
       unstyled: { type: Boolean },
       addCustomSectionIndex: { type: Number, state: true },
@@ -84,9 +89,9 @@ export class EOxStoryTelling extends LitElement {
     /**
      * Enable or disable editor
      *
-     * @type {Boolean}
+     * @type {String}
      */
-    this.showEditor = false;
+    this.showEditor = null;
 
     /**
      * Enable or disable navigation
@@ -197,6 +202,7 @@ export class EOxStoryTelling extends LitElement {
   }
 
   async firstUpdated() {
+    initSavedMarkdown(this);
     addLightBoxScript(this);
 
     // Check if this.#html is initialized, if not, wait for it
@@ -218,6 +224,10 @@ export class EOxStoryTelling extends LitElement {
   }
 
   render() {
+    const editorClass = `${this.showEditor ? "editor-enabled" : ""} editor-${
+      this.showEditor
+    }`;
+
     return html`
       <slot class="slot-hide" @slotchange=${this.handleSlotChange}></slot>
       <style>
@@ -227,13 +237,15 @@ export class EOxStoryTelling extends LitElement {
         ${!this.unstyled && mainStyle}
       </style>
 
-      <div class="story-telling ${this.showEditor ? "editor-enabled" : ""}">
+      <div class="story-telling ${editorClass}">
         <div>${when(this.#html, () => html`${this.#html}`)}</div>
         ${when(
           this.showEditor,
           () => html`
             <eox-storytelling-editor
               .isNavigation=${Boolean(this.showNav)}
+              .storyId=${this.id}
+              show-editor="${this.showEditor}"
               @change=${(e) => {
                 if (e.detail) {
                   this.markdown = e.detail.Story;