+++ elements/storytelling/src/style.eox.js
@@ -160,6 +160,10 @@ ${slider}
     font-size: larger;
     cursor: pointer;
   }
+  .story-telling.editor-enabled.editor-close .section-wrap.section-item::after, 
+  .story-telling.editor-enabled.editor-close .section-wrap.section-item.section-start::before {
+    display: none;
+  }
   .story-telling.editor-enabled .section-wrap.section-item.section-start::before {
     top: -12px;
   }
@@ -252,9 +256,8 @@ ${slider}
   .editor-wrapper {
     z-index: 4;
     width: 35%;
-    height: calc(100vh - 40px);
     position: fixed;
-    top: 20px;
+    bottom: 20px;
     right: 20px;
     border-radius: 10px;
     background: #f2f2f2;
@@ -275,10 +278,6 @@ ${slider}
     cursor: sw-resize;
     z-index: 2;
   }
-  .editor-wrapper.partial-height {
-    height: calc(100vh - 120px);
-    top: 100px
-  }
   #editor {
     width: 100%;
     height: 100%;
@@ -291,8 +290,11 @@ ${slider}
       width: 95%;
     }
   }
-  .editor-hide {
+  .editor-opacity-none {
     opacity: 0;
+  }
+  .editor-hide {
+    display: none;
   }
   .editor-error {
     display: none;
@@ -340,23 +342,12 @@ ${slider}
     margin-left: 0.2rem;
   }
   .editor-saver {
-    width: 22px;
-    height: 22px;
-    border-radius: 50%;
-    position: absolute;
-    top: 10px;
-    left: 10px;
-    animation: rotate 1s linear infinite;
-    display: none;
-  }
-  .editor-saver::before {
-    content: "";
-    box-sizing: border-box;
-    position: absolute;
-    inset: 0px;
-    border-radius: 50%;
-    border: 5px solid #5b5b5b85;
-    animation: spin 2s linear infinite ;
+    position: absolute;    
+    bottom: 18px;
+    left: 17px;
+    font-size: 12px;
+    color: #959694;
+    text-align: right;
   }
   @keyframes rotate {
     100% {transform: rotate(360deg)}