+++ elements/storytelling/test/cases/load-editor.js
@@ -13,7 +13,7 @@ const loadMarkdownEditorTest = () => {
   const jsonFormSelector = `${editorSelector} eox-jsonform`;
 
   cy.mount(
-    `<eox-storytelling show-editor markdown="${testText}"></eox-storytelling>`
+    `<eox-storytelling show-editor="open" markdown="${testText}"></eox-storytelling>`
   ).as(storyTelling);
 
   cy.get(storyTelling)